+++ data/test_instructions/debug.py
@@ -4,12 +4,15 @@
 from pathlib import Path
 from typing import Any
 
-import config_manager
-from beam_calculation.beam_calculator import BeamCalculator
-from beam_calculation.factory import BeamCalculatorsFactory
-from core.accelerator.accelerator import Accelerator
-from core.accelerator.factory import NoFault
-from beam_calculation.simulation_output.simulation_output import SimulationOutput
+from lightwin.beam_calculation.beam_calculator import BeamCalculator
+from lightwin.beam_calculation.factory import BeamCalculatorsFactory
+from lightwin.beam_calculation.simulation_output.simulation_output import (
+    SimulationOutput,
+)
+from lightwin.config import config_manager
+from lightwin.core.accelerator.accelerator import Accelerator
+from lightwin.core.accelerator.factory import NoFault
+
 
 def _set_up_solvers(
     config: dict[str, Any]
@@ -26,7 +29,7 @@ def _set_up_accelerators(
     beam_calculator: BeamCalculator,
 ) -> Accelerator:
     """Create the accelerators."""
-    factory = NoFault(beam_calculator=beam_calculator, **config["files"])
+    factory = NoFault(beam_calculators=beam_calculator, **config)
     accelerator = factory.run()
     return accelerator
 
@@ -38,18 +41,18 @@ def set_up(
     Accelerator,
 ]:
     """Set up everything."""
-    beam_calculator, beam_calculator_ids = _set_up_solvers(config)
+    beam_calculator, _ = _set_up_solvers(config)
     accelerator = _set_up_accelerators(config, beam_calculator)
     return beam_calculator, accelerator
 
 
 def main(
     config: dict[str, dict[str, Any]],
-) -> SimulationOutput:
+) -> tuple[SimulationOutput, Accelerator]:
     """Set up the various faults and fix it."""
     beam_calculator, accelerator = set_up(config)
     simulation_output = beam_calculator.compute(accelerator)
-    return simulation_output
+    return simulation_output, accelerator
 
 
 if __name__ == "__main__":
@@ -57,18 +60,18 @@ if __name__ == "__main__":
     toml_keys = {
         "files": "files",
         # "plots": "plots_minimal",
-        "beam_calculator": "generic_envelope3d",
+        "beam_calculator": "generic_envelope1d",
         "beam": "beam",
     }
     override = {
         "files": {
-            "dat_file": "set_sync_phase.dat",
+            "dat_file": "superpose_map.dat",
         },
-        "beam_calculator": {"n_steps_per_cell": 40},
+        # "beam_calculator": {"n_steps_per_cell": 40},
     }
     config = config_manager.process_config(
         toml_filepath, toml_keys, override=override
     )
-    simulation_output = main(config)
+    simulation_output, accelerator = main(config)
     x = simulation_output.transfer_matrix.individual
     y = simulation_output.transfer_matrix.cumulated