+++ src/lightwin/core/commands/end.py
@@ -2,6 +2,7 @@
 
 from lightwin.core.commands.command import Command
 from lightwin.core.instruction import Instruction
+from lightwin.tracewin_utils.line import DatLine
 
 
 class End(Command):
@@ -10,7 +11,9 @@ class End(Command):
     is_implemented = True
     n_attributes = 0
 
-    def __init__(self, line: list[str], dat_idx: int, **kwargs: str) -> None:
+    def __init__(
+        self, line: DatLine, dat_idx: int | None = None, **kwargs: str
+    ) -> None:
         """Call mother ``__init__``."""
         super().__init__(line, dat_idx)