+++ src/lightwin/core/commands/marker.py
@@ -4,6 +4,7 @@ import logging
 
 from lightwin.core.commands.command import Command
 from lightwin.core.instruction import Instruction
+from lightwin.tracewin_utils.line import DatLine
 
 
 class Marker(Command):
@@ -13,13 +14,12 @@ class Marker(Command):
 
     def __init__(
         self,
-        line: list[str],
-        dat_idx: int,
-        name: str | None = None,
+        line: DatLine,
+        dat_idx: int | None = None,
         **kwargs: str,
     ) -> None:
         """Instantiate the dummy command."""
-        super().__init__(line, dat_idx, name=name)
+        super().__init__(line, dat_idx)
 
     def set_influenced_elements(
         self, instructions: list[Instruction], **kwargs: float