+++ src/lightwin/core/elements/element.py
@@ -14,9 +14,10 @@ import numpy as np
 from lightwin.beam_calculation.parameters.element_parameters import (
     ElementBeamCalculatorParameters,
 )
-from lightwin.core.electric_field import RfField
 from lightwin.core.elements.field_maps.cavity_settings import CavitySettings
+from lightwin.core.em_fields.rf_field import RfField
 from lightwin.core.instruction import Instruction
+from lightwin.tracewin_utils.line import DatLine
 from lightwin.util.helper import recursive_getter, recursive_items
 
 
@@ -47,10 +48,12 @@ class Element(Instruction):
 
     def __init__(
         self,
-        line: list[str],
-        dat_idx: int,
-        name: str | None = None,
-        **kwargs: str,
+        line: DatLine,
+        dat_idx: int | None = None,
+        idx_in_lattice: int = -1,
+        lattice: int = -1,
+        section: int = -1,
+        **kwargs,
     ) -> None:
         """Init parameters common to all elements.
 
@@ -68,24 +71,24 @@ class Element(Instruction):
             later.
 
         """
-        super().__init__(line, dat_idx, name=name)
+        super().__init__(line, dat_idx, **kwargs)
 
         self.elt_info = {
-            "nature": line[0],
+            "nature": line.splitted[0],
         }
-        self.length_m = 1e-3 * float(line[1])
+        self.length_m = 1e-3 * float(line.splitted[1])
 
         # By default, an element is non accelerating and has a dummy
         # accelerating field.
-        self.new_rf_field = RfField()
+        self.rf_field = RfField()
 
         # TODO: init the indexes to -1 or something, to help type hinting
         # dict with pure type: int
         new_idx = {
             "elt_idx": -1,
-            "lattice": -1,
-            "idx_in_lattice": -1,
-            "section": -1,
+            "lattice": lattice,
+            "idx_in_lattice": idx_in_lattice,
+            "section": section,
         }
         self.idx = self.idx | new_idx
         self.beam_calc_param: dict[str, ElementBeamCalculatorParameters] = {}