+++ src/lightwin/core/elements/thin_steering.py
@@ -1,22 +1,22 @@
 """Define :class:`ThinSteering`. It does nothing."""
 
 from lightwin.core.elements.element import Element
+from lightwin.tracewin_utils.line import DatLine
 
 
 class ThinSteering(Element):
     """A dummy object."""
 
+    base_name = "TS"
     increment_lattice_idx = False
     is_implemented = False
 
     def __init__(
         self,
-        line: list[str],
-        dat_idx: int,
-        name: str | None = None,
+        line: DatLine,
+        dat_idx: int | None = None,
         **kwargs: str,
     ) -> None:
         """Force an element with null-length."""
-        super().__init__(line, dat_idx, name)
+        super().__init__(line, dat_idx, **kwargs)
         self.length_m = 0.0
-        self.reinsert_optional_commands_in_line()