+++ tests/test_beam_calculator/test_envelope1d.py
@@ -26,13 +26,30 @@ leapfrog_marker = pytest.mark.xfail(
 )
 
 params = [
-    pytest.param(("RK4", False, False, 40), marks=pytest.mark.smoke),
-    pytest.param(("RK4", False, True, 40), marks=pytest.mark.smoke),
-    pytest.param(("RK4", True, False, 40), marks=pytest.mark.cython),
-    pytest.param(("leapfrog", False, False, 60), marks=leapfrog_marker),
+    pytest.param(
+        ("RK4", False, False, 40),
+        marks=pytest.mark.smoke,
+        id="1D RK4 relative phase",
+    ),
+    pytest.param(
+        ("RK4", False, True, 40),
+        marks=pytest.mark.smoke,
+        id="1D RK4 absolute phase",
+    ),
+    pytest.param(
+        ("RK4", True, False, 40),
+        marks=pytest.mark.cython,
+        id="1D RK4 relative phase Cython",
+    ),
+    pytest.param(
+        ("leapfrog", False, False, 60),
+        marks=leapfrog_marker,
+        id="1D leapfrog relative phase",
+    ),
     pytest.param(
         ("leapfrog", True, False, 60),
         marks=(leapfrog_marker, pytest.mark.cython),
+        id="1D leapfrog relative phase Cython",
     ),
 ]
 
@@ -98,6 +115,7 @@ def simulation_output(
     return my_simulation_output
 
 
+@pytest.mark.envelope1d
 class TestSolver1D:
     """Gater all the tests in a single class."""