+++ tests/test_compensation/test_design_spaces.py
@@ -137,6 +137,7 @@ def simulation_outputs(
     return ref_simulation_output, fix_simulation_output
 
 
+@pytest.mark.envelope1d
 class TestDesignSpaceImplementation:
 
     _w_kin_tol = 1e-3